requirements.txt
aiodns==3.0.0
aiosmtpd==1.4.3
asgiref==3.5.2
atpublic==3.1.1
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
defusedxml==0.7.1
Django==4.1.2
django-ckeditor==6.5.1
django-filter==22.1
django-js-asset==2.0.0
django-rosetta==0.9.8
django-taggit==3.1.0
djangorestframework==3.14.0
idna==3.4
Markdown==3.4.1
oauthlib==3.2.2
Pillow==9.2.0
polib==1.1.1
psycopg2-binary==2.9.5
pycares==4.3.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tzdata==2022.4
urllib3==1.26.13
verify-email==2.4.3